requirements.txt
Django
django-cors-headers
djangorestframework
openai
python-dotenv